requirements.txt
ansimarkup==1.4.0
appdirs==1.4.4
arabic-reshaper==2.1.3
args==0.1.0
asgiref==3.3.4
astroid==2.4.2
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
clint==0.5.1
colorama==0.4.3
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
Django==3.2.3
django-allauth==0.44.0
django-chartjs==2.2.1
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-dbbackup==3.3.0
django-debug-toolbar==3.1.1
django-humanize==0.1.2
django-ipware==3.0.2
django-markdown-deux==1.0.5
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-rest-framework==0.1.0
django-smartfields==1.1.1
django-taggit==1.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fake-useragent==0.1.11
filelock==3.0.12
Fiona @ file:///C:/Users/admin/Downloads/Fiona-1.8.18-cp38-cp38-win32.whl
future==0.18.2
GDAL @ file:///C:/Users/admin/Downloads/GDAL-3.2.1-cp38-cp38-win32.whl
geopandas==0.8.2
gunicorn==20.1.0
html5lib==1.1
httpie==2.4.0
humanize==3.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.5.3
jsonfield==3.1.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3.3
markdown2==2.4.0
matplotlib==3.3.4
mccabe==0.6.1
multitasking==0.0.9
munch==2.5.0
numpy==1.20.1
oauthlib==3.1.0
osrm==0.11.3
pandas==1.2.2
parse==1.19.0
Pillow==8.0.0
polyline==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyee==8.1.0
pyfiglet==0.8.post1
Pygments==2.8.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
pyproj==3.0.0.post1
pyquery==1.4.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.2.0
pytz==2020.1
readtime==1.1.1
reindent==3.5.1
reportlab==3.5.67
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
selenium==3.141.0
service-identity==18.1.0
Shapely==1.7.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
tabula-py==2.2.0
toml==0.10.1
tqdm==4.61.1
txaio==20.4.1
urllib3==1.25.10
virtualenv==20.2.2
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
yfinance==0.1.63
zope.interface==5.2.0
